Use 32-bit signed indirect offset to allow negative offsets

+++ Bit Slicer/ZGSearchData.h
@@ -50,7 +50,7 @@ NS_ASSUME_NONNULL_BEGIN
 	ZGMemorySize _pointerSize;
 	void * _Nullable _rangeValue;
 	double _epsilon;
-	uint16_t _indirectOffset;
+	int32_t _indirectOffset;
 	BOOL _shouldIgnoreStringCase;
 	BOOL _shouldIncludeNullTerminator;
 	
@@ -84,7 +84,7 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic) BOOL shouldCompareStoredValues;
 @property (nonatomic) double epsilon;
 @property (nonatomic) BOOL indirectOffsetMaxComparison;
-@property (nonatomic) uint16_t indirectOffset;
+@property (nonatomic) int32_t indirectOffset;
 @property (nonatomic) uint16_t indirectMaxLevels;
 @property (nonatomic) BOOL indirectStopAtStaticAddresses;
 @property (nonatomic) BOOL filterHeapAndStackData;
